code_v2/requirements.txt
pandas == 0.24.1
pytorch == 1.7.1
numpy == 1.16.2
scipy == 1.1.0
scikit-learn == 0.23.2